core_service/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-axes==6.5.0
django-environ==0.11.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
redis==5.0.7
celery==5.4.0
django-celery-beat==2.6.0
pytest==8.2.2
pytest-django==4.8.0
pytest-asyncio==0.23.7
factory-boy==3.3.0
